move and rename shared combobox hooks
